requirements.txt
croniter==2.0.1
PyYAML==6.0.1
watchdog==3.0.0
requests==2.31.0
pytz==2023.3.post1
fastapi==0.104.1
uvicorn==0.24.0.post1
Jinja2==3.1.2
apscheduler==3.10.4
websockets==12.0
pypinyin==0.50.0
bcrypt==4.0.1
passlib==1.7.4
python-multipart==0.0.6
cn2an==0.5.22
pystray==0.19.5